+++ src/routes/MainRoutes.jsx
@@ -3,6 +3,8 @@ import Footer from "../components/8-footer/Footer";
 import Navbar from "../components/0-navbar/Navbar";
 import ErrorScreen from "../pages/ErrorScreen";
 import ProScreen from "../pages/ProScreen";
+import CoachScreen from "../pages/CoachScreen";
+import LegalWarningScreen from "../pages/LegalWarningScreen";
 
 const MainRoutes = () => {
   return (
@@ -10,6 +12,10 @@ const MainRoutes = () => {
       <Navbar />
       <Routes>
         <Route path="/" element={<ProScreen />} />
+        <Route path="/training" element={<CoachScreen />} />
+        <Route path="/privacy" element={<LegalWarningScreen type="privacy" />} />
+        <Route path="/cookies" element={<LegalWarningScreen type="cookies" />} />
+        <Route path="/legal" element={<LegalWarningScreen type="legal" />} />
         <Route path="*" element={<ErrorScreen />} />
       </Routes>
       <Footer />